HabitService/requirements.txt
fastapi
uvicorn 
pydantic 
python-jose
sqlalchemy
psycopg2
psycopg2-binary
